010301a03: uber bug & segfaults fixed
I've finally resolved a very long standing issue that has been causing may people, including my self, to experience sporadic segfaults.  Usually these would out put a 'pthread->mutex', 'pa_thread_mutex', or etc error.  This was because get2gnow's notifications were attempting to play the default sound, just a window 'pop' sound.  But currently this causes a pulseaudio threading issue resulting in a segfault. Well: now its no longer messing with me.

I've also added support to get2gnow's notification popup's so they now display the user's avatar along with their update.

I have several other bug fixes and improvement implemented as well.  I'm doing some major refactoring to get2gnow that will vastly improve get2gnow's preformance.  But only part of this has been implemented... along with my never ending list of new features, further tests, and as always: eventual bug fixes.  Despite this being only a partial update fixing several possible segfauls, especially the pulseaudio segfault bug fix: \*w00t!\*.

Committed on : 2010-03-01 @ 03:11:53 AM

+++ src/gconfig.c
@@ -81,7 +81,7 @@ typedef struct {
 	gint		cached_value_int;
 	
 	gchar		*cached_key_float;
-	gdouble		cached_value_float;
+	gfloat		cached_value_float;
 	
 	gchar		*cached_key_string;
 	gchar		*cached_value_string;
@@ -124,7 +124,7 @@ static void gconfig_finalize(GObject *object);
 
 static gboolean gconfig_check_cached_bool(const gchar *key, gboolean *value);
 static gboolean gconfig_check_cached_int(const gchar *key, gint *value);
-static gboolean gconfig_check_cached_float(const gchar *key, gdouble *value);
+static gboolean gconfig_check_cached_float(const gchar *key, gfloat *value);
 static gboolean gconfig_check_cached_string(const gchar *key, gchar **value);
 static gboolean gconfig_check_cached_list(const gchar *key, GSList **value);
 static gboolean gconfig_check_cached(GConfigSupportedValue which_cache, const gchar *key);
@@ -297,7 +297,7 @@ gboolean gconfig_get_int(const gchar *key, gint *value){
 	return TRUE;
 }/*gconfig_get_int(key, &int);*/
 
-gboolean gconfig_set_float(const gchar *key, gdouble value){
+gboolean gconfig_set_float(const gchar *key, gfloat value){
 	if(G_STR_N_EMPTY(gconfig_priv->cached_key_float) && g_str_equal(gconfig_priv->cached_key_float, key) )
 		gconfig_priv->cached_value_float=value;
 	
@@ -312,7 +312,7 @@ gboolean gconfig_set_float(const gchar *key, gdouble value){
 	return success;
 }/*gconfig_set_float(key, float);*/
 
-gboolean gconfig_get_float(const gchar *key, gdouble *value){
+gboolean gconfig_get_float(const gchar *key, gfloat *value){
 	if(gconfig_check_cached_float(key, value))
 		return TRUE;
 	
@@ -542,12 +542,12 @@ static gboolean gconfig_check_cached_int(const gchar *key, gint *value){
 	return TRUE;
 }/*gconfig_check_cached_int(const gchar *key, gint *value);*/
 
-static gboolean gconfig_check_cached_float(const gchar *key, gdouble *value){
+static gboolean gconfig_check_cached_float(const gchar *key, gfloat *value){
 	if(!gconfig_check_cached(GConfigFloat, key))
 		return FALSE;
 	*value=gconfig_priv->cached_value_float;
 	return TRUE;
-}/*gconfig_check_cached_float(const gchar *key, gdouble *value);*/
+}/*gconfig_check_cached_float(const gchar *key, gfloat *value);*/
 
 static gboolean gconfig_check_cached_string(const gchar *key, gchar **value){
 	if(!gconfig_check_cached(GConfigString, key))
